fix react key issue and refactoring
