+++ README.md
@@ -5,7 +5,8 @@ Prototype of Likemark client from scratch with Webpack + React + MobX + Semantic
 **Be careful**: all guidelines are describded in the collaborative [Team Chiru Guidelines Draft](https://paper.dropbox.com/doc/Team-Chiru-Guidelines-Draft-YHh8yPhFpgz4b5cRPXlgA) which needs to be read before starting development. All suggestions are welcomed!
 
 1. Install `yarn` with your favorite package manager : `apt install yarn`, `brew install yarn`, `choco install yarn` or `npm install yarn --global`
+2. Clone the git rpository and checkout the **dev** branch
 2. Get all dependencies with `yarn install`
 3. Start the **webpack server** with `yarn start` or build it with `yarn build`
-4. You should debug the project with the associated **Visual Studio Code** configuration.
+4. You should debug the project with the associated **Visual Studio Code** configuration
