+++ src/app.js
@@ -2,8 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import { Sidebar, Menu, Icon } from 'semantic-ui-react'
 
-import { LikemarkList } from './components/LikemarkList'
-import { LikemarkStore } from './observables/LikemarkStore'
+import { LikemarkStore, LikemarkListView } from './components/LikemarkList'
 
 import 'semantic-ui-sidebar/sidebar.min.css'
 import 'semantic-ui-segment/segment.min.css'
@@ -15,7 +14,7 @@ const store = new LikemarkStore()
 
 ReactDOM.render(
   <div>
-    <LikemarkList store={store} />
+    <LikemarkListView store={store} />
     <Sidebar as={Menu} direction='bottom' visible>
       <Menu.Item name='add' onClick={store.addLikemark.bind(store)}>
         <Icon name='add circle' />
