+++ src/components/Likemark.js
@@ -1,14 +1,22 @@
 import React from 'react'
-
+import * as UUIDv4 from 'uuid/v4'
 import { List } from 'semantic-ui-react'
 
 import 'semantic-ui-list/list.min.css'
 import 'semantic-ui-icon/icon.min.css'
 
-export class LikemarkView extends React.Component {
+class LikemarkItem {
+  constructor (name, description) {
+    this.id = UUIDv4()
+    this.name = name || ''
+    this.description = description || ''
+  }
+}
+
+class LikemarkView extends React.Component {
   render () {
     return (
-      <List.Item>
+      <List.Item key={this.props.model.id}>
         <List.Icon name='bookmark' />
         <List.Content>
           <List.Header>{this.props.model.name}</List.Header>
@@ -18,3 +26,5 @@ export class LikemarkView extends React.Component {
     )
   }
 }
+
+export { LikemarkItem, LikemarkView }
