--- src/observables/Likemark.js
@@ -1,6 +1,0 @@
-export class Likemark {
-  constructor (name, description) {
-    this.name = name || ''
-    this.description = description || ''
-  }
-}
